One good implementation of BR in the inactive set.
The new call is
  br_inactive_mark(u, lb, residual, theta=0.8)
It uses Stefano's implementation where the residual is defined
by the user.  Note the returned error estimate is only valid
if the residual comes from Poisson.

This should be used with caution if the residual is *not* from
Poisson because the scaling of the two terms in the B&R formula
is likely to be wrong.

This version also has better documentation.