Isolating and documenting the various limitations.
The most important change is that the not-at-all-basic
parallel testing has been moved out of tests/test_basic.py
into tests/test_parallel.py.  In the new file the
function names will not be found by pytest's default algorithm,
which is deliberate so that `pytest .` actually works
on standard installs.  See the end of README.md for manual
testing of parallel code.

Add explicit setting of distribution parameters to 2
examples, and switch one to be a udomark() example.

Document limitations in README.md.

Also cleaning out old comments and run black in
the tests/ directory.