Change vcdmark() solver to fixed iterations of CG+ICC
The solver is 4 fixed iterations of ICC-preconditioned
CG.  In parallel the preconditioner is ASM with ICC on
the blocks.  Testing shows that the new solver is 3 to 10
times faster on fine meshes.  (But note that the VCD
solve is quick in any case compared to the main VI
solver.)

There are small effects on the final meshes from VCD,
but looking at examples seems to show that the new-solver
meshes are no better or worse.  For low resolutions the
meshes seem to just come out the same anyway.

Options allow reverting to the default (direct; LU by
MUMPS) solver, and to print solver timing.