Clarify VCD algorithm.
One change is not-a-change.  The time dependent strong form
of the backward Euler diffusion heat equation step is in fact
  u - u0
  ------ = Div(D grad u)
    1
This is what we are already using in VCD.  This has the weak
form used currently if D = 0.5 h^2.  Note this D is spatially
variable so it matters whether it is inside the divergence or
not.  This will be documented in the paper.

This is different from
  u - u0
  ------ = Div(grad u)
    dt
where dt = 0.5 h^2, which has a different weak form from
what is currently in the code.

Two actual changes:
  1. UFL expression "h = CellDiameter(mesh)" is used in
  weak form, so that there is no need to store "timestep",
  which is now gone.  This saves a bit of memory but causes
  slight changes in high-resolution meshes.
  2. The ends of the bracket are now consistent with the
  figure in the paper.  Their meaning is reversed by this
  change.

Also ran black on some stuff.